refactor: extract helper functions
